+++ source/subtitle_sync/changelog.md
@@ -1,4 +1,9 @@
 
+**<span style="color:#56adda">0.0.4</span>**
+- install ffs with init.d script instead of requirements.txt file
+- fix language comparator to use iso639 match to handle cases where configured
+  language and actual subtitle language are a mix of 2 and 3 letters
+
 **<span style="color:#56adda">0.0.3</span>**
 - fix progress parser
 - fix type variable name to not override builtin function